ai-professor-vs/rag_server/requirements.txt
fastapi
uvicorn
qdrant-client
llama-index
llama-index-llms-groq
llama-index-vector-stores-qdrant
llama-index-embeddings-huggingface
sentence-transformers
python-dotenv
